Added meshbits electrumx servers for kmd and beer.
These have been tested a bit and have got monitoring in place at 5 minute interval.
These are automated and are using Floating/Elastic IPs so we can keep the IP consistent even if we were to change underlying instance.
Once we're happy with the stability of the service, we'll give it a DNS name.